Refactor members models into multiple files
