+++ docs/members.models.rst
@@ -1,0 +1,43 @@
+members.models package
+======================
+
+.. automodule:: members.models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+members.models.email\_change module
+-----------------------------------
+
+.. automodule:: members.models.email_change
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+members.models.member module
+----------------------------
+
+.. automodule:: members.models.member
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+members.models.membership module
+--------------------------------
+
+.. automodule:: members.models.membership
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+members.models.profile module
+-----------------------------
+
+.. automodule:: members.models.profile
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
