+++ website/members/models/__init__.py
@@ -1,0 +1,4 @@
+from .member import *
+from .profile import *
+from .membership import *
+from .email_change import *
